Pass signals received by syslog onto handler subprocess
When syslog is feeding its output to a subprocess and receives a SIGHUP,
SIGINT, SIGUSR1 or SIGUSR2, pass this signal on to the subprocess. In
addition, exit cleanly on receipt of SIGTERM, leaving the subprocess to
finish cleanly too.

Although we use SA_RESTART which avoids EINTR from read() and recv(), we
need to recover from an interrupted poll() correctly.