Add a trigger option to ueventd
If ueventd is run with -t, a change event is synthesized for all
registered modules, buses and devices so that ueventd rules can be
applied to them where necessary on system cold-plug. This option is
deliberately not retained when the daemon is restarted following a
SIGHUP, so the restart won't result in a replay of the cold-plug.