Avoid holding /dev/console open in PID 1
On Linux, SAK or SysRq-k kills all processes controlled by the current
tty or holding it open. The init process is not exempt, so Linux systems
unceremoniously crash if SAK is used on the console while init has it open.

To guard against this, close all file descriptors on startup and reopen
/dev/console when required for running the startup and shutdown scripts or a
single-user shell.