Fix /bin/stop to fork before calling reboot(RB_POWER_OFF)
reboot() never returns when called with RB_HALT_SYSTEM or RB_AUTOBOOT,
but RB_POWER_OFF has odd asynchronous behaviour and does return. The
safest way to use it is apparently to fork and call it from the child,
pausing the parent to avoid a premature exit.

/bin/stop is safe to exec directly at the end of /etc/init with this
change, so fix examples/etc/init to illustrate this.