Ensure syslog handles time zones sensibly on musl and glibc
syslog(3) in both glibc and musl sends traditional BSD-format datagrams
to /dev/log. The dates sent by glibc are in the time zone of the calling
process, whereas on musl systems, these time stamps are always UTC.

Select the right behaviour automatically, depending on the C library,
but allow an explicit override by defining UTCLOG as 0 or 1.

Default to emitting UTC log lines unless compiled with -DUTC=0. Local
time is ambiguous in time zones where daylight saving changes occur, so
this probably isn't a sensible configuration.

On musl systems, running a -DUTC=0 syslog with TZ=UTC in the environment
will correctly override the output zone from local time to UTC at run
time, but this won't work on glibc as it also affects the time zone used
for datagram parsing.