Add seal utility to exec a program from a sealed memfd
This provides an easy way to protect an arbitrary program from attack via
the strange Linux /proc/self/exe magic when run from the host but joining
a potentially hostile container.

  seal PROG [ARG]...

locates PROG on the PATH, clones it to a new sealed memfd, then executes
the memfd with the given arguments using fexecve(). If PROG contains '/',
it is directly used as the source pathname, bypassing PATH.

Compare the output of readlink /proc/self/exe executed directly:

  $ readlink /proc/self/exe
  /bin/readlink

with the output when it is run from a sealed memfd:

  $ seal readlink /proc/self/exe
  /memfd:/bin/readlink (deleted)