Add reap: a completely null init
This can be convenient for PID 1 in a PID namespace, where nothing but
a pure process reaper is required.

We explicitly set the disposition of SIGCHLD to SIG_IGN before calling
pause(). POSIX defines the resulting behaviour: the exit status of
children is cleanly discarded and they do not become zombies.

Another option for a minimal init is to loop on wait(), but this will
spin in a busy loop when it is the only process and has no children.