Take precautions against uevent buffer overflow
recv() on the uevent socket fails with ENOBUFS if events are read too
slowly so the socket receive buffer fills up. This can happen at
startup if events are retriggered for every device in a large system
with slow ueventd.conf handler functions.

Enlarge the receive buffer sufficiently that it should never overflow
even if output is completely blocked and a large number of events are
retriggered by ueventd -t.

If uevent suffers ENOBUFS despite this, report it as a special overflow
action rather than a fatal error.