Add '-s PATH' and '-t HOST:PORT' socket server options to daemon
Incorporate a subset of traditional inetd or tcpserver functionality
into daemon.

TCP and unix stream addresses can be specified with -t HOST:PORT and
-s PATH respectively. daemon will listen on sockets bound to these
addresses and run the specified command as a handler for each inbound
connection.

Where a hostname provided to -t resolves to a number of IPv4/IPv6
addresses, listening sockets will be bound to each of them. HOST can
also be an IPv4 literal, an IPv6 literal in square brackets, or be
omitted completely to bind a wildcard socket for the specified port.