Configure syslog UTC mode at runtime
Use local time in log entries only if TZ is explicitly set and non-empty
in the environment. Otherwise, default to UTC.

Suppress an overflow warning for using tm_gmtoff of struct tm in an int
context. Its value is bounded between -86400 and +86400 by the length
of a day, and int is always 32-bit or larger on Linux.