Remove subprocess support from uevent
It is no harder to run uevent as a child of ueventd rather than as the
parent, and this allows both a simpler uevent.c and a more logical
process tree.

Take care to kill the child correctly when terminating, but we keep it
alive across a SIGHUP-triggered re-exec to avoid dropping uevents.