Remove subprocess support from syslog
It is no harder to run syslog as a child of syslogd rather than as the
parent, and this allows both a simpler syslog.c and a more logical
process tree.

Take care to kill the child correctly when terminating, but we keep it
alive across a SIGHUP-triggered re-exec to avoid dropping log messages.