Work around a bash idiosyncrasy in the example /etc/init
When bash is blocked on read, it still handles SIGCHLD and waits for
dead children to avoid zombies. However, when a timeout is requested,
bash won't wait on children until the read returns. This leads to a
short-lived zombie from process substitution in the watchdog daemon.

Explicitly wait for <(:) to terminate before using read -t N to sleep.