Use nodev, noexec and nosuid in example /etc/fstab
Where appropriate, add the nodev, noexec and nosuid mount options to
filesystems in the example /etc/fstab.

Device nodes aren't required outside /dev. Nothing needs to be executable
except in / and /tmp, and we can forbid setuid and setgid binaries in /tmp.
A redundant nosuid option is not required where noexec is already in force.

Add the memory_recursiveprot and nsdelegate options to /sys/fs/cgroup as
these are now the standard configuration.

Drop the explicit relatime from / because this is now enabled by default.
Use PARTLABEL=root with auto type rather than a fixed ext4 /dev/vda rootfs,
and add a similar /boot mount with PARTLABEL=boot.