Add optional double fork in daemon
By default, daemon intentionally runs the command as a session and
process group leader. On Linux, a session leader without a controlling
terminal can acquire one just by opening a terminal device, unlike BSD.
Sometimes this behaviour can be inconvenient, so add a -f flag to fork
twice so we no longer lead the session.

This repurposes a flag previously ignored for command-line compatibility
with the original BSD daemon(8), but in practice the options we support
already collide with more modern BSD versions.