Explain why we can safely strncpy() into struct sockaddr_un
On Linux, sun_path in struct sockaddr is NUL-padded not NUL-terminated:
the pathname is implicitly terminated by the end of the fixed-size array.
If we insist on a terminating NUL, we can't bind valid full-length
socket paths. The strncpy() in listen_unix() is thus correct even though
it appears suspicious. Explicitly mention this in a comment.