Add ueventwait to wait for a single device
A common boot-time pattern is to wait for a single device which is
asynchronously discovered. It is wasteful to run a persistent ueventd
just to cover this case.

Provide a ueventwait script which waits for a single device, matching
pattern arguments of the form KEY=PATTERN against uevent properties of
existing devices then newly-added devices until a match is found.