Keep syslogd in the foreground until /dev/log exists
This eliminates any risk of a race between syslogd creating /dev/log
and subsequently-started daemons calling syslog() to write to it, which
might lead to the loss of early log messages. Exit with an error if the
background daemon doesn't create /dev/log within two seconds, to avoid
stalling boot indefinitely.