Don't fork daemon solely to unlink a pidfile on exit
When we listen on a socket or supervise a command, we fork and run the
command in a child. If a pidfile is specified, the PID of the parent
daemon process is stored in it and it is unlinked if the parent exits.

If -p is given without -r, -s or -t, lock and write the pidfile, then
directly exec the command with the file descriptor held open, instead of
retaining a parent process just to unlink a pidfile when its child exits.
The pidfile will remain locked until the child exits unless it closes it.

Although this assumes good behaviour from the command, unsupervised -p
is a cleaner, more ergonomic building block to integrate into scripts.